move package test to package_test.go
